refactor urlsession
