+++ Tests/Fakes/FakeURLSessionDataTask.swift
@@ -1,14 +1,25 @@
-//import Foundation
-//
-///// Fakes a URLSessionDataTask for testing.
-//class FakeURLSessionDataTask: URLSessionDataTask {
-//    private let closure: () -> Void
-//
-//    init(closure: @escaping () -> Void) {
-//        self.closure = closure
-//    }
-//
-//    override func resume() {
-//        closure()
-//    }
-//}
+import Foundation
+
+/// A fake data task that will immediately call completion.
+public class FakeURLSessionDataTask: URLSessionDataTask {
+  private let urlResponse: URLResponse?
+  private let data: Data?
+  private let fakedError: Error?
+  private let completionHandler: (Data?, URLResponse?, Error?) -> Void
+
+  public init(
+    urlResponse: URLResponse?,
+    data: Data?,
+    error: Error?,
+    completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void
+  ) {
+    self.urlResponse = urlResponse
+    self.data = data
+    self.fakedError = error
+    self.completionHandler = completionHandler
+  }
+
+  public override func resume() {
+    completionHandler(data, urlResponse, fakedError)
+  }
+}
