+++ Tests/PayID/PayIDClientTest.swift
@@ -1,60 +1,6 @@
 import XCTest
 import XpringKit
 
-/// A fake URLSession that will return data tasks which will call completion handlers with the given parameters.
-public class FakeURLSession: URLSession {
-  public var urlResponse: URLResponse?
-  public var data: Data?
-  public var error: Error?
-
-  public convenience init(string: String, urlResponse: HTTPURLResponse?, error: Error?) {
-    let data = string.data(using: .utf8)
-    self.init(data: data, urlResponse: urlResponse, error: error)
-  }
-
-  public init(data: Data?, urlResponse: HTTPURLResponse?, error: Error?) {
-    self.data = data
-    self.urlResponse = urlResponse
-    self.error = error
-  }
-
-  public override func dataTask(
-    with request: URLRequest,
-    completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void
-  ) -> URLSessionDataTask {
-    return FakeURLSessionDataTask(
-      urlResponse: urlResponse,
-      data: data,
-      error: error,
-      completionHandler: completionHandler
-    )
-  }
-}
-
-/// A fake data task that will immediately call completion.
-public class FakeURLSessionDataTask: URLSessionDataTask {
-  private let urlResponse: URLResponse?
-  private let data: Data?
-  private let fakedError: Error?
-  private let completionHandler: (Data?, URLResponse?, Error?) -> Void
-
-  public init(
-    urlResponse: URLResponse?,
-    data: Data?,
-    error: Error?,
-    completionHandler: @escaping (Data?, URLResponse?, Error?) -> Void
-    ) {
-    self.urlResponse = urlResponse
-    self.data = data
-    self.fakedError = error
-    self.completionHandler = completionHandler
-  }
-
-  public override func resume() {
-    completionHandler(data, urlResponse, fakedError)
-  }
-}
-
 class PayIDClientTest: XCTestCase {
 
   func testXRPAddressForInvalidPayID() {
